+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Make formula page improvements
 - Fix incorrect stock update when making formula
 - Formula is now automatically marked as made and under evaluation status when making is complete
+- Add option to reset ingredient quantity to its original value when making the formula
+- Detect and notify if an ingredient is overdosed when making a formula
+- Rewrite sell formula page
 
 ### Version 5.7
 - A rating option has been added so each formula can be rated