+++ releasenotes.md
@@ -2,4 +2,6 @@ Whats New:
 ----------------------
 - Make formula page improvements
 - Formula is now automatically marked as made and under evaluation status when making is complete
-
+- Add option to reset ingredient quantity to its original value when making the formula
+- Detect and notify if an ingredient is overdosed when making a formula
+- Rewrite sell formula page